Bug 30317: Replace the use of jQueryUI tabs on search history page
This patch replaces jQueryUI tabs on the staff interface's search
history page, replacing them with Bootstrap tabs.

To test, apply the patch and perform some bibliographic and
authority searches in the staff client if necessary.

Go to the search history page and test that the "Catalog" and
"Authority" tabs work correctly.

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>