Bug 30401: Convert budgets administration page tabs to Bootstrap
This patch updates the budgets administration page to use Bootstrap tabs
instead of jQueryUI.

This patch contains indentation changes, so ignore whitespace when you
use your preferred diff tool.

To test, apply the patch and go to Administration -> Budgets. Check that
the "Active budgets" and "Inactive budgets" work correctly.

Append "?tab=2" to the page URL to confirm that the second tab is
preselected.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>