Bug 30419: Convert authority detail page tabs to Bootstrap
This patch updates the authority detail page to use Bootstrap tabs
instead of jQueryUI.

The patch also removes some broken JavaScript. It will be re-implemented
in Bug 30422.

To test, apply the patch search for an authority record in the
authorities module.

View the detail page for the authority and test that the tabs look
correct and work correctly.

Signed-off-by: Séverine Queune <severine.queune@bulac.fr>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>