Bug 30424: Reindent advanced search template in the staff interface
This patch reindents the advanced search template. Some lines have been
split to make the markup more readable.

To test, apply the patch and test the advanced search page in the staff
interface. Everything should work as normal. Be sure to test both the
"More options" and "Fewer options" views.

When viewing the diff for this change, ignore whitespace. The only
visible changes should be split lines.

Signed-off-by: Séverine QUEUNE <severine.queune@bulac.fr>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>