Bug 29155: (follow-up) Work around change to jQueryUI tab API
For some reason, with the combination of jQuery 3.6.0 and jQueryUI
1.13.1, the way to access the current active tab's link has changed.
Previously the tab activate event would provide information about the
link as "ui.newTab.context." It appears that "context" is no longer
available.

This patch updates the two instances in Koha where this process is
broken.

To test, apply the patch and check out to a patron with holds on their
account.

- When the checkout page loads, confirm that each tab loads its contents
  correctly when clicked.
- Check that the correct tab is activated when you append the tab link
  to the page url, e.g.

  /cgi-bin/koha/circ/circulation.pl?borrowernumber=XXX#reldebarments

  ...which should select the "Restrictions" tab for you. You may need to
  try this in a new tab or shift-reload to get the browser to look for
  the hash.

- Perform the same tests on the patron details page.

Signed-off-by: David Cook <dcook@prosentient.com.au>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>